The geometry package will use float instead of double, though the raytracer will still use double.

+++ Rings/src/com/almostrealism/geometry/Scaled.java
@@ -1,7 +1,7 @@
 package com.almostrealism.geometry;
 
 public interface Scaled {
-	public void setScale(double x, double y, double z);
+	public void setScale(float x, float y, float z);
 	
 	public float[] getScale();
 }
